+++ src/main/kotlin/com/mojh/cms/security/jwt/TokenStatus.kt
@@ -1,4 +1,4 @@
-package com.mojh.cms.security
+package com.mojh.cms.security.jwt
 
 enum class TokenStatus {
     VALID, INVALID, EXPIRED
